wip: refactoring camToken
